requirements.txt
# ===== Web Framework =====
fastapi==0.110.0
uvicorn[standard]==0.27.1

# ===== Data Validation / Settings =====
pydantic==2.6.1
python-dotenv==1.0.1

# ===== HTTP / Networking =====
requests==2.31.0
aiohttp==3.9.3

# ===== Data Processing =====
numpy==1.26.4
pandas==2.2.0

# ===== Utility =====
tqdm==4.66.1
python-dateutil==2.8.2
